Replicates rekordbox's beatjump shift mode
Pressing shift + pad 8 increases the jump on all other buttons by a factor of 16
(up to a maximum of 128 beats backwards & forwards on buttons 7 & 8 respectively).

Pressing shift + pad 7 decreases the jump on all other buttons by a factor of 16
(down to a minimum of 1/16 beats backwards & forwards on buttons 1 & 2 respectively).