Simplifies beatjump pad lighting
Relies solely on <outputs> for lighting up the required pads in
beatjump mode.

This removes the need to track the controllers padmode in the script.

This is desirable, because the controller tracks the padmode in its
firmware and seems to be susceptible to benign but unexpected
behavior when that process is interfered with (e.g., manually
resetting the padmode lights sometimes prevents the next pad mode
from being enabled until a sequence of other buttons are pressed).